@rfourquet 2018-03-21
> the specialized code for sets/dicts when origin and destination are different might not be worth adding.

I would say, now that it's implemented, we can keep it. Maybe for other types of `Abstract{Set,Dict}` it can make a substantial difference.