@rfourquet 2018-03-21
I see that in you benchmark code, you didn't interpolate the global variables `A`, `B`, `sA` and `sB`, could you check if it makes a difference in this case?