@rfourquet 2018-03-23
I worked on this and will push a commit to feed the discussion (I'm not yet fully happy with it).

I'm still reluctant at calling `replace!` a method which has an implicit precondition (like `count >= 0`), i.e. which is not meant to be used by end-users.

One thing I noticed also is that with the new method, there is an ambiguity in a call like `replace!(x->2x, [1], [2])`: the `replace!(pred, A, new)` is called... so we should either get rid of this one, or find a new name for the one taking both `res` and `A`. I would now favor not making it public for now, as it would not be available yet for other forms of `replace!`, i.e. the ones with `pred` and pairs; I called it `replaceimpl!` temporarily in the commit I will push.

Also, `Nothing` is needed for type-stability only when the replacement is specified with pairs, for `replace`. What do you think that we limit the use of a `count=nothing` default  only for this method; then when calling the underlying `replace!` method, we pass `count=coalesce(count, typemax(Int))`. For the two other advantages of using `nothing` that you mentioned: maybe constant propagation can work? and SIMD could still be worked out with a more complicated code for the implementation for vectors. For me it would be best if people have simply to implement `replaceimpl!(new, res, A, count::Integer=typemax(Int))`; it just needs a test for `count <= 0`; it doesn't prevent them from re-using the already implemented code when their type is `Union{AbstractArray,AbstractDict,AbstractSet}`, by implementing directly the `_replace!` method, like `Base` does. Alternatively, (which I do in my commit, and apparently doesn't alter the performance of your version, except for the set `sA`, not sure yet why), the default implementation `replaceimpl!` itself transforms `typemax(Int)` to `nothing` for the cases where it matters.