@rfourquet 2018-03-23
My pushed commit may not seem to improve so much, but I guess my main points are: 1) don't call `replace!` a non-API method,  as it can confuse the user and it makes it more clear also for the developer when reading the code (at least in my view), and 2) if `nothing` is useful for performance, use it only where necessary, but don't pollute all the methods (and the API) with it.