@nalimilan 2018-03-24
Actually, `count=nothing` isn't strictly needed for SIMD, as we can simply check `count >= length(A)` and use two different loops (sometimes LLVM can do this automatically via loop unswitching, but here it doesn't seem to work; not such about constant propagation). So I've removed all uses of `count=nothing` except for the `replace` method with pairs. I've also added a `check_count` method which takes care of converting to `Int` and throwing an error if negative. All of this allowed getting rid of `replaceimpl`, which I find easy to confuse with `_replace!` (the jungle of methods is already hard to grasp). Finally, I've renamed the `_replace!` method taking pairs to `replace_pairs`, since it's really different from others.

Let's say we keep `_replace!` for now, meaning that there are no API stability guarantees regarding how to implement `replace` and `replace!` (other than by implementing the public methods directly).