@rfourquet 2018-03-27
Looks good! The use of `nothing` in the implementation for arrays seemed to allow a shorter code but I prefer your new version. I can review more closely tomorrow, but I think this will be good to go.