@rfourquet 2018-03-28
Would be good indeed if someone comments on the "singleton stuff", basically that `replace([1, missing], missing=>2) isa Vector{Int}` (when `count` is not specified).