@nalimilan 2018-04-11
Travis failure on Mac is unrelated. I'll merge tomorrow barring objections.